fix date filter

+++ EDocument_Data/Models/Shared/JwtSettings.cs
@@ -4,7 +4,7 @@
     {
         public required string Issuer { get; init; }
         public required string Audience { get; init; }
-        public required int DurationInDays { get; init; }
+        public required double DurationInDays { get; init; }
 
     }
